+++ lib/solid_objects/effect_executor.rb
@@ -113,6 +113,7 @@ module SolidObjects
 
     # @rbs () -> Effect?
     def claim_next
+      EffectRecoveryCoordinator.new.recover_available
       database_adapter.transaction do
         now = database_adapter.database_now
         effect = database_adapter.lock_candidates(
@@ -178,6 +179,7 @@ module SolidObjects
       )
       result_message = nil
       database_adapter.transaction do
+        Instance.lock.find(effect.instance_id)
         locked_effect = Effect.lock.find(effect.id)
         verify_claim!(locked_effect)
         result_message = enqueue_result_message(
@@ -213,6 +215,7 @@ module SolidObjects
     def fail_effect(effect, error)
       result_message = nil
       database_adapter.transaction do
+        Instance.lock.find(effect.instance_id)
         locked_effect = Effect.lock.find(effect.id)
         verify_claim!(locked_effect)
         dead = locked_effect.attempt_count >= locked_effect.max_attempts