+++ lib/solid_objects/test_helper.rb
@@ -35,6 +35,7 @@ module SolidObjects
           ClaimedMessage,
           ReadyMessage,
           Broadcast,
+          EffectRecovery,
           Effect,
           Reminder,
           Message,